Add attribute proxy for transition between plaintext and ciphertext fields
vault_attribute_proxy defines proxy getter and setter methods

It overrides the getter and setter for a particular non-encrypted attribute so that they also call the getter/setter of the encrypted one.
This ensures that all the code that uses the attribute in question also updates/retrieves the encrypted value whenever it is available.

This method is useful if you have a plaintext attribute that you want to replace with a vault attribute.
During a transition period both attributes can be seamlessly read/changed at the same time.
Then by using the boolean option encrypted_attribute_only, you will be able to test if the ciphertext field works as expected before getting rid of the plaintext attribute.