Add a serializer for IPAddr objects
In postgresql there are inet and cidr types for columns that are handled by
the `ActiveRecord::ConnectionAdapters::PostgreSQL::OID::Cidr` and
`ActiveRecord::ConnectionAdapters::PostgreSQL::OID::Inet` type classes.
This means we get ruby `IPAddr` objects as the value for our attributes
when we read them out of the database with `ActiveRecord`.  We want to
support these column types out of the box so we've introduced an `ipaddr`
serializer that knows how to handle `IPAddr` objects (the implementation of
which is based heavily on the postgresql adapter type class mentioned
above).